+++ src/rdkafka_assignor.h
@@ -137,7 +137,7 @@ rd_kafka_assignor_run (struct rd_kafka_cgrp_s *rkcg,
 rd_kafka_assignor_t *
 rd_kafka_assignor_find (rd_kafka_t *rk, const char *protocol);
 
-int rd_kafka_assignors_init (rd_kafka_t *rk, char *errstr, int errstr_size);
+int rd_kafka_assignors_init (rd_kafka_t *rk, char *errstr, size_t errstr_size);
 void rd_kafka_assignors_term (rd_kafka_t *rk);
